Use optional chaining to add listeners to optional UI elements.
Because some UI elements are optional, code assuming they always exist
was causing some of the JS to fail.

Notably, the mobile menu was inoperable unless the '...' sidebar was
enabled, since trying to add a listener to the non-existent sidebar
button would error out the JS function before it could add the listener
for the mobile menu.

Same changes included in the minified js.